requirements.txt
wikipedia==1.4.0
python-telegram-bot==21.1.1
requests==2.31.0
